+++ lib/kinematics/kinematics.h
@@ -5,4 +5,4 @@
 
 float standarize_deg(int32_t lon_or_lat);
 float calculate_azimuth(float tracker_x, float tracker_y, float object_x, float object_y);
-float optimize_azimuth(float current_deg, float destination_deg);
+int optimize_azimuth(int current_deg, int destination_deg);
